refactor: CommentException 위치 수정 (post -> comment)
