+++ src/main/java/until/the/eternity/dcs/domain/comment/application/CommentService.java
@@ -12,8 +12,8 @@ import until.the.eternity.dcs.domain.comment.dto.response.CommentPageResponseIte
 import until.the.eternity.dcs.domain.comment.dto.response.CommentPersistResponse;
 import until.the.eternity.dcs.domain.comment.entity.Comment;
 import until.the.eternity.dcs.domain.comment.entity.CommentRepository;
-import until.the.eternity.dcs.domain.post.exception.CommentModifyForbiddenException;
-import until.the.eternity.dcs.domain.post.exception.CommentNotFoundException;
+import until.the.eternity.dcs.domain.comment.exception.CommentModifyForbiddenException;
+import until.the.eternity.dcs.domain.comment.exception.CommentNotFoundException;
 import until.the.eternity.dcs.domain.user.application.UserService;
 import until.the.eternity.dcs.domain.user.entity.UserSummary;
