+++ src/main/java/until/the/eternity/dcs/domain/comment/exception/CommentModifyForbiddenException.java
@@ -1,8 +1,8 @@
-package until.the.eternity.dcs.domain.post.exception;
+package until.the.eternity.dcs.domain.comment.exception;
 
 import until.the.eternity.dcs.common.exception.CustomException;
 
-import static until.the.eternity.dcs.domain.post.exception.CommentExceptionCode.COMMENT_MODIFY_FORBIDDEN_EXCEPTION;
+import static until.the.eternity.dcs.domain.comment.exception.CommentExceptionCode.COMMENT_MODIFY_FORBIDDEN_EXCEPTION;
 
 public class CommentModifyForbiddenException extends CustomException {
 	public CommentModifyForbiddenException() {
