+++ src/main/java/until/the/eternity/dcs/domain/comment/exception/CommentNotFoundException.java
@@ -1,8 +1,8 @@
-package until.the.eternity.dcs.domain.post.exception;
+package until.the.eternity.dcs.domain.comment.exception;
 
 import until.the.eternity.dcs.common.exception.CustomException;
 
-import static until.the.eternity.dcs.domain.post.exception.CommentExceptionCode.COMMENT_NOT_FOUND_EXCEPTION;
+import static until.the.eternity.dcs.domain.comment.exception.CommentExceptionCode.COMMENT_NOT_FOUND_EXCEPTION;
 
 public class CommentNotFoundException extends CustomException {
 	public CommentNotFoundException(Long id) {
