+++ src/test/java/until/the/eternity/dcs/domain/comment/application/CommentServiceTest.java
@@ -12,7 +12,7 @@ import until.the.eternity.dcs.domain.comment.dto.response.CommentPageResponseIte
 import until.the.eternity.dcs.domain.comment.dto.response.CommentPersistResponse;
 import until.the.eternity.dcs.domain.comment.entity.Comment;
 import until.the.eternity.dcs.domain.post.entity.Post;
-import until.the.eternity.dcs.domain.post.exception.CommentNotFoundException;
+import until.the.eternity.dcs.domain.comment.exception.CommentNotFoundException;
 import until.the.eternity.dcs.domain.user.application.UserService;
 import until.the.eternity.dcs.domain.user.fake.FakeUserService;
 import until.the.eternity.dcs.infrastructure.FakeCommentRepository;
